Models class created
